feat: minor changes
Ref. #35
