+++ app/src/main/res/drawable/pen_solid.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="32dp"
+    android:height="32dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M362.7,19.3L314.3,67.7 444.3,197.7l48.4,-48.4c25,-25 25,-65.5 0,-90.5L453.3,19.3c-25,-25 -65.5,-25 -90.5,0zM291.7,90.3L58.6,323.5c-10.4,10.4 -18,23.3 -22.2,37.4L1,481.2C-1.5,489.7 0.8,498.8 7,505s15.3,8.5 23.7,6.1l120.3,-35.4c14.1,-4.2 27,-11.8 37.4,-22.2L421.7,220.3 291.7,90.3z"/>
+</vector>
