+++ app/src/main/res/layout/fragment_restaurant_page.xml
@@ -105,6 +105,19 @@
             app:layout_constraintRight_toRightOf="parent"
             tools:ignore="HardcodedText" />
 
+        <ImageButton
+            android:id="@+id/restaurant_post_comment"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="20dp"
+            android:background="@drawable/round"
+            android:contentDescription="Open Google Maps"
+            android:padding="10dp"
+            android:src="@drawable/pen_solid"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintLeft_toLeftOf="parent"
+            tools:ignore="HardcodedText" />
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </layout>
